demo 0.0.8

+++ rt-thread/bsp/hc32f460/applications/app_key.h
@@ -7,7 +7,7 @@ typedef struct{
 	uint8_t		Brush_key_Flag;		//盘刷电机按键
 	uint8_t		All_key_Flag;		//盘刷+吸风电机按键
 	uint8_t		Off_key_Flag;		//OFF按键
-	int16_t		Hand_FLag;			//把手开关
+	uint8_t		Hand_FLag;			//把手开关
 	uint8_t		Work_Status;		//工作状态
 }TKeyInfoDef;
 
@@ -15,7 +15,7 @@ typedef struct {
 	uint8_t		Suction_key_Flag;	//吸风电机按键
 	uint8_t		Brush_key_Flag;		//盘刷电机按键
 	uint8_t		All_key_Flag;		//盘刷+吸风电机按键
-	int16_t		Hand_FLag;			//把手开关
+	uint8_t		Hand_FLag;			//把手开关
 }TKey_CountInfoDef;
